[IMP] website: share carousel horizontal controllers
This commit shares and adapts the style and options of the horizontal
controllers used in `s_carousel_intro` so that it can be reused in other
carousel snippets.

task-4094405
Part of task-4077427
