+++ addons/website/views/snippets/s_carousel_intro.xml
@@ -56,8 +56,8 @@
                     </div>
                 </div>
             </div>
-            <div class="s_carousel_intro_controllers container position-absolute start-50 bottom-0 translate-middle-x w-100 mb-md-4 o_not_editable">
-                <div class="s_carousel_intro_controllers_row row gap-3 gap-lg-5 justify-content-between flex-nowrap flex-row-reverse">
+            <div class="o_horizontal_controllers container position-absolute start-50 bottom-0 translate-middle-x w-100 mb-md-4 o_not_editable">
+                <div class="o_horizontal_controllers_row row gap-3 gap-lg-5 justify-content-between flex-nowrap flex-row-reverse mx-0">
                     <!-- Controls -->
                     <div class="o_arrows_wrapper gap-2 w-auto p-0">
                         <button class="carousel-control-prev o_not_editable" contenteditable="false" t-attf-data-bs-target="#myCarouselIntro{{uniq}}" data-bs-slide="prev" aria-label="Previous" title="Previous">
@@ -79,35 +79,6 @@
             </div>
         </div>
     </section>
-</template>
-
-<template id="snippet_options_carousel_intro" inherit_id="website.snippet_options_carousel" primary="True">
-    <xpath expr="//div" position="attributes">
-        <attribute name="data-target">.s_carousel_intro</attribute>
-        <attribute name="data-exclude"/>
-        <attribute name="data-js">CarouselIntro</attribute>
-    </xpath>
-    <xpath expr="//we-button[@data-select-class='s_carousel_arrows_hidden']" position="attributes">
-        <attribute name="data-name">carousel_arrows_hidden_opt</attribute>
-        <attribute name="data-hidden-controllers">true</attribute>
-    </xpath>
-    <xpath expr="//we-button[@data-select-class='s_carousel_indicators_hidden']" position="attributes">
-        <attribute name="data-name">carousel_indicators_hidden_opt</attribute>
-        <attribute name="data-hidden-controllers">true</attribute>
-    </xpath>
-    <xpath expr="//we-select" position="replace">
-        <we-select string="Controllers" data-apply-to=".s_carousel_intro_controllers_row">
-            <we-button data-select-class="justify-content-between">Default</we-button>
-            <we-button data-select-class="justify-content-between flex-row-reverse">Reversed</we-button>
-            <we-button data-select-class="justify-content-center" data-name="carousel_controllers_centered_opt"  data-dependencies="carousel_arrows_hidden_opt, carousel_indicators_hidden_opt">Centered</we-button>
-        </we-select>
-    </xpath>
-</template>
-
-<template id="s_carousel_intro_options" inherit_id="website.snippet_options">
-    <xpath expr="//t[@t-call='website.snippet_options_background_options']" position="after">
-        <t t-call="website.snippet_options_carousel_intro"/>
-    </xpath>
 </template>
 
 <record id="website.s_carousel_intro_000_scss" model="ir.asset">
