+++ addons/website/views/snippets/snippets.xml
@@ -623,6 +623,27 @@
     </div>
 </template>
 
+<template id="snippet_options_carousel_bottom_controllers" inherit_id="website.snippet_options_carousel" primary="True">
+    <xpath expr="//div" position="attributes">
+        <attribute name="data-target">.s_carousel_intro</attribute>
+        <attribute name="data-exclude"/>
+        <attribute name="data-js">Carousel</attribute>
+    </xpath>
+    <xpath expr="//we-button[@data-select-class='s_carousel_arrows_hidden']" position="replace">
+        <we-button data-select-class="s_carousel_arrows_hidden" data-name="carousel_arrows_hidden_opt" data-toggle-controllers="">Hidden</we-button>
+    </xpath>
+    <xpath expr="//we-button[@data-select-class='s_carousel_indicators_hidden']" position="replace">
+        <we-button data-select-class="s_carousel_indicators_hidden" data-name="carousel_indicators_hidden_opt" data-toggle-controllers="">Hidden</we-button>
+    </xpath>
+    <xpath expr="//we-select" position="replace">
+        <we-select string="Controllers" data-apply-to=".o_horizontal_controllers_row">
+            <we-button data-select-class="justify-content-between">Default</we-button>
+            <we-button data-select-class="justify-content-between flex-row-reverse">Reversed</we-button>
+            <we-button data-select-class="justify-content-center" data-name="carousel_controllers_centered_opt"  data-dependencies="carousel_arrows_hidden_opt, carousel_indicators_hidden_opt">Centered</we-button>
+        </we-select>
+    </xpath>
+</template>
+
 <template id="snippet_options_border_line_widgets">
     <we-row t-att-string="label">
         <we-input data-name="border_width_opt"
@@ -896,6 +917,7 @@
     </t>
 
     <t t-call="website.snippet_options_carousel"/>
+    <t t-call="website.snippet_options_carousel_bottom_controllers"/>
 
     <div data-js="CarouselItem"
          data-selector=".s_carousel .carousel-item, .s_quotes_carousel .carousel-item, .s_carousel_intro .carousel-item">
